Moved stuff around and added middleware to require application/json on all requests
