+++ index.js
@@ -1,5 +1,16 @@
 var mongoose = require('mongoose');
-var api = require('./routes.js');
+var express = require('express');
+var app = express();
+var bodyParser = require('body-parser');
+app.use(bodyParser.json());
+
+// Add middleware
+var middleware = require('./middleware/middleware');
+middleware(app)
+
+// Add routes
+var routes = require('./routes/routes');
+routes(app)
 
 mongoose.connect('mongodb://localhost/test');
 var db = mongoose.connection;
@@ -8,7 +19,7 @@ db.once('open', function() {
     console.log("Connected to db")
 
     // Start web server
-    api.listen(3000, function() {
+    app.listen(3000, function() {
         console.log("Server started");
     });
 });
