+++ middleware/middleware.js
@@ -1,0 +1,3 @@
+module.exports = function(app) {
+    require('./headers')(app)
+}
