+++ routes/routes.js
@@ -1,0 +1,4 @@
+
+module.exports = function(app) {
+    require('./persons')(app)    
+}
